Add "reattach-to-user-namespace" command to the beginning of tmux.conf
WHY:

I noticed that in tmux sessions certain commands didn't work as
expected. In particular, passing a file or directory name to 'sublime,'
which should open Sublime with that file or directory, was just
opening Sublime as if I had used the command with no arg.

I found this solution on StackOverflow. It uses the
tmux-MacOSX-pasteboard package, which I had already
installed for other config.